LW-40: Reorgaize packages to separate tests that they dont use Spring Context
