+++ backend/src/main/java/com/letswork/api/app/shared/ParameterizedException.java
@@ -1,0 +1,8 @@
+package com.letswork.api.app.shared;
+
+public abstract class ParameterizedException extends RuntimeException {
+
+    public ParameterizedException(String message) {
+        super(message);
+    }
+}
