+++ backend/src/main/java/com/letswork/api/app/token/domain/EmailSender.java
@@ -1,0 +1,8 @@
+package com.letswork.api.app.token.domain;
+
+import javax.mail.MessagingException;
+
+interface EmailSender {
+
+    void sendEmail(String to, String confirmationToken) throws MessagingException;
+}
