+++ backend/src/main/java/com/letswork/api/app/token/domain/TokenFactory.java
@@ -1,0 +1,17 @@
+package com.letswork.api.app.token.domain;
+
+import com.letswork.api.app.user.domain.UserEntity;
+
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
+class TokenFactory {
+
+    TokenEntity create(UserEntity userEntity) {
+        return TokenEntity.builder()
+                .confirmationToken(UUID.randomUUID().toString())
+                .dateInSeconds(String.valueOf(TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis())))
+                .user(userEntity)
+                .build();
+    }
+}
