+++ backend/src/main/java/com/letswork/api/app/user/domain/UserFactory.java
@@ -1,0 +1,14 @@
+package com.letswork.api.app.user.domain;
+
+import com.letswork.api.app.user.domain.dto.CreateUserDto;
+
+class UserFactory {
+
+    UserEntity create(CreateUserDto dto) {
+        return UserEntity.builder()
+                .email(dto.getEmail())
+                .password(dto.getPassword())
+                .isEnabled(false)
+                .build();
+    }
+}
