--- backend/src/main/java/com/letswork/api/shared/BaseEntity.java
@@ -1,19 +1,0 @@
-package com.letswork.api.shared;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-@MappedSuperclass
-@Getter
-@Setter
-public abstract class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
