--- backend/src/main/java/com/letswork/api/shared/ParameterizedException.java
@@ -1,8 +1,0 @@
-package com.letswork.api.shared;
-
-public abstract class ParameterizedException extends RuntimeException {
-
-    public ParameterizedException(String message) {
-        super(message);
-    }
-}
