--- backend/src/main/java/com/letswork/api/token/domain/EmailSender.java
@@ -1,8 +1,0 @@
-package com.letswork.api.token.domain;
-
-import javax.mail.MessagingException;
-
-interface EmailSender {
-
-    void sendEmail(String to, String confirmationToken) throws MessagingException;
-}
