--- backend/src/main/java/com/letswork/api/token/domain/TokenFactory.java
@@ -1,17 +1,0 @@
-package com.letswork.api.token.domain;
-
-import com.letswork.api.user.domain.UserEntity;
-
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-
-class TokenFactory {
-
-    TokenEntity create(UserEntity userEntity) {
-        return TokenEntity.builder()
-                .confirmationToken(UUID.randomUUID().toString())
-                .dateInSeconds(String.valueOf(TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis())))
-                .user(userEntity)
-                .build();
-    }
-}
