--- backend/src/main/java/com/letswork/api/token/domain/TokenRepository.java
@@ -1,16 +1,0 @@
-package com.letswork.api.token.domain;
-
-import org.springframework.data.repository.Repository;
-
-import java.util.Set;
-
-interface TokenRepository extends Repository<TokenEntity, Long> {
-
-    void save(TokenEntity token);
-
-    void delete(TokenEntity token);
-
-    Set<TokenEntity> findAll();
-
-    TokenEntity findByConfirmationToken(String confirmationToken);
-}
