--- backend/src/main/java/com/letswork/api/user/domain/UserFactory.java
@@ -1,14 +1,0 @@
-package com.letswork.api.user.domain;
-
-import com.letswork.api.user.domain.dto.CreateUserDto;
-
-class UserFactory {
-
-    UserEntity create(CreateUserDto dto) {
-        return UserEntity.builder()
-                .email(dto.getEmail())
-                .password(dto.getPassword())
-                .isEnabled(false)
-                .build();
-    }
-}
