--- backend/src/main/java/com/letswork/api/user/domain/dto/CreateUserDto.java
@@ -1,13 +1,0 @@
-package com.letswork.api.user.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CreateUserDto {
-
-    private String email;
-    private String password;
-    private String confirmedPassword;
-}
